Add Get Grid helper
This takes a collection of lines and turns it into a grid -- multi-dimensional array where each character corresponds to an x, y.

It might be nice to expand this out into a class at some point that could be reused. Maybe composed of cells within a grid.

It seems problems often have scenarios where you need to access neighbors which could be accomplished in a grid ;)