Add Generics to Grid and Add StringGrid
StringGrid is a default implementation, otherwise a mapper can now be provided to convert the strings into the generic type.

This will allow for custom types such as enums, etc to be utilized to convert a character into an enum value. This will allow for easier lookup and identification in code.