Add optional Idx to mapFind
This allows us to act on the index while mapping.

I tried to add an optional index to the find method, but wasn't certain if this is even possible due to the generator that comes back from the map. #antiPattern.

Hopefully calling `entries` on the map collection is not too expensive 😬 maybe there's a cheaper way for this? #anotherDaysProblem.