Solve Day 10 Part 1
Maybe overkill with the object oriented approach.

Definitely plenty of room for improvement. I believe there's some jank going on with the going vs coming from logic and inverting directions that is likely not needed. It might just require remapping some of the lookups as well. It might be nice to switch some of the lookups to Maps, but that too may be overkill.

The underlying code takes an OO approach, while the implementation is more functional.

I should have some logger class that I can turn off and on. The only issue is that I often only want to log for the current tests I'm working on. Maybe each logger could be an instance and I could turn on logging based off of instances/keys. Rather than leaving logging as comments it could be part of the code.