Flatten Day 03 Folder Structure
Since the inputs and examples are reused, it kind of seems silly to have them in different directories.

If this ends up being the standard the plop file should be updated to reflect this, so that there's more consistency moving forward.

I also like the pattern of having a utility/service, so I should also consider that as well.