Extract Day 03 Number Filtering Logic
This will allow us to reuse this logic for a new gear lookup that will check to see if a number is by a gear and find all the numbers next to the gear.

This won't be super efficient, but it should work out just fine.

This will have us checking all the numbers for each gear when we could just utilize the valid number set and assign them to the nearby gears for them.