Add Day 04 Scored Card
This is a card with a score on it. Previously in Part 1 all this information was being flattened since it was extraneous, also building the score into the card was a bit janky since it was taking a card object to compute the score. That could have been refactored, but instead I decided to create a new type that extended the initial base card type.