Day 05 - Move Instruction Set Up
Originally I thought I would have some files that the instruction set would consume but didn't want to really expose.

I was going to utilize dart convention by sticking them into a directory and prefixing them with `_` on the files that shouldn't be exposed upwards, but since this is not part of the JS language feature set, I decided to just keep it vanilla and not export the functions and stick them all in a single file, while testing it via the exposed function, rather than testing the portion directly.