Add Day 05 - Seeds Part 2
This follows the new part 2 rules, and now with the new seeds we should be able to push those into our part 1 solving code.

My concern is that the collection will likely be HUGE. At that point I'm not certain how to proceed if that is a problem. Given the large numbers I think the seed collection length is going to become very large. If it solves in a semi-reasonable time then I may disable this test after it solves. It's also possible something like dynamic programming or memoization is required here.